+++ integrations/aws-v3/aws/core/helpers/types.py
@@ -25,8 +25,10 @@ class ObjectKind(StrEnum):
     CODEBUILD_BUILD_RUN = "AWS::CodeBuild::BuildRun"
     CODEDEPLOY_APPLICATION = "AWS::CodeDeploy::Application"
     CODEDEPLOY_DEPLOYMENT_GROUP = "AWS::CodeDeploy::DeploymentGroup"
+    CODEDEPLOY_DEPLOYMENT = "AWS::CodeDeploy::Deployment"
     CODEDEPLOY_DEPLOYMENT_TARGET = "AWS::CodeDeploy::DeploymentTarget"
     CODEPIPELINE_PIPELINE = "AWS::CodePipeline::Pipeline"
+    CODEPIPELINE_STAGE = "AWS::CodePipeline::Stage"
     CODEPIPELINE_ACTION = "AWS::CodePipeline::Action"